// dist/commonjs/renderer/shader/variable/EVariableType.d.ts
export declare enum EVariableType {
    FLOAT_1,
    FLOAT_2,
    FLOAT_3,
    FLOAT_4,
    VECTOR_2,
    VECTOR_3,
    VECTOR_4,
    COLOR_3,
    FLOAT_MAT3,
    FLOAT_MAT4,
    BUFFER,
    SAMPLER_CUBE,
    SAMPLER_2D,
    NUMBER_1,
    STRUCTURE,
    STRUCTURES,
    SAMPLER_ARRAY,
    FLOAT_MAT4_ARRAY,
}